+++ src/features/helper/types/diagnosticSetting.ts
@@ -1,0 +1,4 @@
+export interface DiagnosticSetting {
+	rule: string
+	severity: string
+}