+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to the "sqlfluff" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.1.5] - 2022-10-22
+
+- Add `linter.severity` configuration option.
+
 ## [0.1.4] - 2022-10-11
 
-- Add `format.arguments` and `linter.arguments`.
+- Add `format.arguments` and `linter.arguments` configuration options..
 
 ## [0.1.3] - 2022-10-02